feat(gui): 2D map export overhaul, viewer visibility controls and settings polish
Reworks the world-map export path (plan, manifest ownership, exporter) and the 3D
viewer's export button, and adds viewer visibility controls: selected-reference and
collision visibility, projection controls, a draw-distance tooltip converter, an
in-game teleport clipboard command, and renderer-backend labelling. Settings-panel
organization and the accessibility ratchet are updated to match.

+++ src/BethesdaMultitool/App/Controls/WorldView3D/WorldView3DControl.ExportPanel.cs
@@ -163,7 +163,7 @@ public sealed partial class WorldView3DControl
             _exportScale, tiled, Export3DMaxTileDimension, Export3DMaxImageDimension);
 
         var pxPerCell = (int)Math.Round(MapExport3DPlanner.BaseScalePxPerUnit * _exportScale * 4096f);
-        var totalTiles = plan.Cols * plan.Rows;
+        var totalTiles = (long)plan.Cols * plan.Rows;
         string tileNote;
         if (tiled && totalTiles > 1)
         {
@@ -395,7 +395,7 @@ public sealed partial class WorldView3DControl
 
         // Per-tile PNGs + manifest only when the user asked for tiles AND there is more than one; a
         // one-tile "tiled" run stays a single PNG. A non-tiled grid > 1 tile is stitched into one PNG.
-        var tiledOutput = tiledRequested && (plan.Cols * plan.Rows) > 1;
+        var tiledOutput = tiledRequested && ((long)plan.Cols * plan.Rows) > 1;
 
         // World-space centers of the worldspace's ACTUAL cells — lets the tile loop skip a provably
         // empty tile before spending a full GPU render + readback (see RunProjectionExportAsync).
